lang_graph_agent/pyproject.toml
[project]
name = "lang-graph"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.59.0",
    "json-repair>=0.55.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
]
